5885: accept fees modal rename
