+++ src/components/accept-fees-modal/AcceptFeesModal.tsx
